+++ marginaleffects/__init__.py
@@ -1,5 +1,5 @@
 from .comparisons import avg_comparisons, comparisons
-from .datagrid import datagrid
+from .datagrid import datagrid, datagridcf
 from .hypotheses import hypotheses
 from .predictions import avg_predictions, predictions
 from .slopes import avg_slopes, slopes